@@ -4,7 +4,8 @@
   "description": "Hallelujah! This Angular module offers a HAL/JSON http-client to easily interact with a Spring Data Rest API or any API that implements the Spring Data Rest resource model",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "packagr": "ng-packagr -p ng-package.json"
+    "packagr": "ng-packagr -p ng-package.json",
+    "lint": "tslint src/**/*.ts public_api.ts"
   },
   "keywords": [
     "Spring Data Rest",
@@ -15,23 +16,37 @@
     "Angular 2"
   ],
   "author": "unibit",
+  "contributors": ["cgatian"],
+  "maintainers": ["chrismarth", "pinksummit"],
   "license": "MIT",
   "homepage": "https://github.com/pinksummit/HALlelujah",
   "issues": {
-    "url": "https://github.com/pinksummit/HALlelujah/issues"
+    "url": "https://pinksummit.atlassian.net/"
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/pinksummit/HALlelujah"
   },
   "devDependencies": {
-    "ng-packagr": "^1.3.0",
-    "typescript": "2.3.4"
+    "@angular/common": "^10.0.0",
+    "@angular/core": "^10.0.0",
+    "@angular/compiler": "^10.0.0",
+    "@angular/compiler-cli": "^10.0.0",
+    "codelyzer": "^6.0.0",
+    "ng-packagr": "^10.0.0",
+    "rxjs": "~6.6.0",
+    "tslint": "^6.1.2",
+    "typescript": "~3.9.7",
+    "zone.js": "^0.10.3"
+  },
+  "dependencies": {
+    "tslib": "^2.0.0"
   },
   "peerDependencies": {
-    "@angular/common": ">=5.0.0",
-    "@angular/core": ">=5.0.0",
-    "rxjs": ">=5.4.3",
-    "zone.js": "^0.8.18"
+    "@angular/common": "^10.0.0",
+    "@angular/core": "^10.0.0",
+    "rxjs": "~6.6.0",
+    "zone.js": "^0.10.3",
+    "typescript": "~3.9.7"
   }
 }
