@@ -1,7 +1,5 @@
-import {Observable} from 'rxjs/Observable';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/catch';
-import {HttpClient, HttpHeaders, HttpParams} from '@angular/common/http';
+import {Observable} from 'rxjs';
+import {HttpClient} from '@angular/common/http';
 import {EntityHelper} from './entity-helper';
 import {RelatedCollectionRef, RelatedEntityRef} from './references';
 
