@@ -1,7 +1,4 @@
-/* SystemJS module definition */
-
-
-export * from "./src/hallelujah.module";
+export * from './src/hallelujah.module';
 export * from './src/api.service';
 export * from './src/entity';
 export * from './src/entity-helper';
