@@ -2,6 +2,7 @@ import {HttpClient, HttpParams} from '@angular/common/http';
 import {Entity} from './entity';
 import {Page} from './page';
 
+// @dynamic
 export class EntityHelper {
 
   static transformForUpdate(object: Object): Object {
@@ -16,7 +17,7 @@ export class EntityHelper {
     return result as Object;
   }
 
-  static initEntity<E extends Entity>(type: { new(): E }, payload: Object, http: HttpClient): E {
+  static initEntity<E extends Entity>(type: new() => E, payload: Object, http: HttpClient): E {
     const entity = new type();
     for (const p in payload) {
       entity[p] = payload[p];
@@ -25,13 +26,12 @@ export class EntityHelper {
     return entity;
   }
 
-  static initEntityCollection<E extends Entity>(type: { new(): E }, payload: any, http: HttpClient): E[] {
+  static initEntityCollection<E extends Entity>(type: new() => E, payload: any, http: HttpClient): E[] {
     return payload._embedded[Object.keys(payload['_embedded'])[0]]
       .map(item => EntityHelper.initEntity(type, item, http));
-
   }
 
-  static initPage<E extends Entity>(type: { new(): E }, payload: any, http: HttpClient): Page<E> {
+  static initPage<E extends Entity>(type: new() => E, payload: any, http: HttpClient): Page<E> {
     const page = new Page(type, http);
     page.items = EntityHelper.initEntityCollection(type, payload, http);
     page.totalItems = payload.page ? payload.page.totalElements : page.items.length;
